% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bevstat.R
\name{calculate_age}
\alias{calculate_age}
\title{calculating age}
\usage{
calculate_age()
}
\value{
two tables, one for age distribution and one for mean age
}
\description{
calculates the relative age for each commune and outputs it in a table in the global environment. Also gives the mean age.
}
