% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durchschnittsalter.R
\name{durchschnittsalter}
\alias{durchschnittsalter}
\title{Title}
\usage{
durchschnittsalter(my_data)
}
\arguments{
\item{my_data}{daten ausgelesen mit \code{\link{read_stat_data}}}
}
\value{

}
\description{
Title
}
