% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-population-data.R
\name{get_population_data}
\alias{get_population_data}
\title{Download Statistik Austria data}
\usage{
get_population_data()
}
\value{
A data frame with 392508 rows and 7 variables:
\describe{
  \item{time_section}{In this case just the year 2020}
  \item{sex}{male or female}
  \item{state_number}{A number between 1 and 9 indicating one of the states in Austria}
  \item{commune_number}{A 5 digit number encoding the commune}
  \item{commune}{The name of the commune}
  \item{age}{The corresponding age}
  \item{number}{The number of people with corresponding age, sex and living in the specified commune}
}
}
\description{
Download the population data
from Statistik Austria
\href{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_bevstandjbab2002_BevStand_2020}{here}.
The raw data is downloaded and a single data frame is returned.
The data frame contains the columns time_section, sex, state_number, commune_number, commune, age, number.
}
\examples{
\dontrun{
population_df <- get_population_data()
}
}
