% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-some-lit-scatterplot.R
\name{plot_preschool_vs_average_age}
\alias{plot_preschool_vs_average_age}
\title{Plot average age per commune vs the proportion of preschool kids}
\usage{
plot_preschool_vs_average_age(population_summary_df, alpha = 0.5)
}
\arguments{
\item{population_summary_df}{The data frame returned by \link[bevstat]{summarise_population_df}}

\item{alpha}{opacity}
}
\value{
A scatter plot of average age per commune vs the proportion of preschool kids
}
\description{
Plot average age per commune vs the proportion of preschool kids
}
