% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-population-data.R
\docType{data}
\name{population_df}
\alias{population_df}
\title{Population data of Austria in the year 2020}
\format{
A data frame with 392508 rows and 7 variables:
\describe{
  \item{time_section}{In this case just the year 2020}
  \item{sex}{male or female}
  \item{state_number}{A number between 1 and 9 indicating one of the states in Austria}
  \item{commune_number}{A 5 digit number encoding the commune}
  \item{commune}{The name of the commune, note that communes are not unique, i.e., Krumbach!!}
  \item{age}{The corresponding age}
  \item{number}{The number of people with corresponding age, sex and living in the specified commune}
}
}
\source{
\url{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_bevstandjbab2002_BevStand_2020}
}
\usage{
population_df
}
\description{
A dataset containing the population data of Austria
in the year 2020.
This data set can also be downloaded via by \link[bevstat]{get_population_data}.
}
\keyword{datasets}
