% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bevstat.R
\name{print_row}
\alias{print_row}
\title{prints row in age distribution table}
\usage{
print_row(y)
}
\arguments{
\item{y}{}
}
\value{
relative age distribution and mean age
}
\description{
writting the "katestral" code of the commune into the function will output the relative age distribution as well as the mean age in the commune
}
