% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bevstat.R
\name{read_bevstat}
\alias{read_bevstat}
\title{copies finished Dataset in global environment}
\usage{
read_bevstat()
}
\value{
A Dataframe which is in working condition
}
\description{
This function downloads the Data.frame from the URL: https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_bevstandjbab2002_BevStand_2020.
The names of the collumns are automatically assigned. Additionally all codes are replaced with their corresponding labels.
The Data types are set, so they can be used for statistical analysis
}
