% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stat_data.R
\name{read_stat_data}
\alias{read_stat_data}
\title{Auslesen der Bevölkerungsdaten}
\usage{
read_stat_data(x)
}
\arguments{
\item{x}{url als Charakterstring}
}
\value{
Datensatz mit Labels und Headern
}
\description{
Auslesen der Bevölkerungsdaten
}
