% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-data.R
\name{summarise_population_df}
\alias{summarise_population_df}
\title{Calculate summary statistics of population}
\usage{
summarise_population_df(df)
}
\arguments{
\item{df}{The data frame returned by \link[bevstat]{get_population_data} or \link[bevstat]{population_df} directly.}
}
\value{
A data frame with 6 variables:
\describe{
  \item{state}{e.g., "Wien", "Burgenland" etc.}
  \item{commune_number}{A 5 digit number encoding the commune}
  \item{commune}{The name of the commune}
  \item{relative_size_age_group_per_commune}{The relative size of each age group per commun}
  \item{average_age_per_commune}{Average age in each commune}
}
}
\description{
Takes the data frame returned by \link[bevstat]{get_population_data} or \link[bevstat]{population_df} directly
and calculates the relative size of each age group per commune as well as the average age in each commune.
Note that the information regarding gender is first summarized.
Furthermore, the state name is added, e.g., "Wien", "Burgenland" etc.
}
